requirements.txt
videosdk-agents
videosdk
videosdk-plugins-google
python-dotenv